Refractoring code.
