+++ react_native/styles/image.tsx
@@ -1,4 +1,4 @@
 export const pin = {
-    width: 20,
-    height: 20,
+    width: 30.5,
+    height: 35,     // ratio: 1.170731707317 x (men då försvann lite av höger sida så fick tweeka lite)
 }
